Create Answers for SELECT from Nobel Tutorial
Understanding the question in a logical way that was the hardest thing for me to learn SQL by myself for the first time. The only way to solve that problem which is trying to practice as much as you can, although the process might be difficult at the very beginning.